refactor ConsulServiceDiscovery into separate struct
